Capstone/requirements.txt
-i https://pypi.org/simple
absl-py==2.1.0; python_version >= '3.7'
asttokens==3.0.0; python_version >= '3.8'
astunparse==1.6.3
blinker==1.9.0; python_version >= '3.9'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.8; python_version >= '3.7'
comm==0.2.2; python_version >= '3.8'
contourpy==1.3.1; python_version >= '3.10'
cycler==0.12.1; python_version >= '3.8'
debugpy==1.8.9; python_version >= '3.8'
decorator==5.1.1; python_version >= '3.5'
distlib==0.3.9
executing==2.1.0; python_version >= '3.8'
filelock==3.17.0; python_version >= '3.9'
flask==3.1.0; python_version >= '3.9'
flatbuffers==24.3.25
fonttools==4.55.1; python_version >= '3.8'
fsspec==2025.2.0; python_version >= '3.8'
gast==0.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-pasta==0.2.0
grpcio==1.68.1; python_version >= '3.8'
h5py==3.12.1; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
ipykernel==6.29.5; python_version >= '3.8'
ipython==8.30.0; python_version >= '3.10'
itsdangerous==2.2.0; python_version >= '3.8'
jedi==0.19.2; python_version >= '3.6'
jinja2==3.1.5; python_version >= '3.7'
joblib==1.4.2; python_version >= '3.8'
jupyter-client==8.6.3; python_version >= '3.8'
jupyter-core==5.7.2; python_version >= '3.8'
keras==3.7.0; python_version >= '3.9'
kiwisolver==1.4.7; python_version >= '3.8'
libclang==18.1.1
markdown==3.7; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
matplotlib==3.9.3; python_version >= '3.9'
matplotlib-inline==0.1.7; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
ml-dtypes==0.4.1; python_version >= '3.9'
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0; python_version >= '3.5'
networkx==3.4.2; python_version >= '3.10'
numpy==1.26.4; python_version >= '3.9'
nvidia-cublas-cu12==12.4.5.8; python_version >= '3'
nvidia-cuda-cupti-cu12==12.4.127; python_version >= '3'
nvidia-cuda-nvcc-cu12==12.5.82; python_version >= '3'
nvidia-cuda-nvrtc-cu12==12.4.127; python_version >= '3'
nvidia-cuda-runtime-cu12==12.4.127; python_version >= '3'
nvidia-cudnn-cu12==9.1.0.70; python_version >= '3'
nvidia-cufft-cu12==11.2.1.3; python_version >= '3'
nvidia-curand-cu12==10.3.5.147; python_version >= '3'
nvidia-cusolver-cu12==11.6.1.9; python_version >= '3'
nvidia-cusparse-cu12==12.3.1.170; python_version >= '3'
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5; python_version >= '3'
nvidia-nvjitlink-cu12==12.4.127; python_version >= '3'
nvidia-nvtx-cu12==12.4.127; python_version >= '3'
opt-einsum==3.4.0; python_version >= '3.8'
optree==0.13.1; python_version >= '3.7'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
parso==0.8.4; python_version >= '3.6'
pexpect==4.9.0
pillow==11.0.0; python_version >= '3.9'
pipenv==2024.4.1; python_version >= '3.8'
platformdirs==4.3.6; python_version >= '3.8'
prompt-toolkit==3.0.48; python_full_version >= '3.7.0'
protobuf==4.25.5; python_version >= '3.8'
psutil==6.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.18.0; python_version >= '3.8'
pyparsing==3.2.0; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2024.2
pyzmq==26.2.0; python_version >= '3.7'
requests==2.32.3; python_version >= '3.8'
rich==13.9.4; python_full_version >= '3.8.0'
scikit-learn==1.6.1; python_version >= '3.9'
scipy==1.14.1; python_version >= '3.10'
seaborn==0.13.2; python_version >= '3.8'
setuptools==75.8.0; python_version >= '3.9'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
stack-data==0.6.3
sympy==1.13.1; python_version >= '3.8'
tensorboard==2.18.0; python_version >= '3.9'
tensorboard-data-server==0.7.2; python_version >= '3.7'
tensorflow==2.18.0; python_version >= '3.9'
termcolor==2.5.0; python_version >= '3.9'
threadpoolctl==3.5.0; python_version >= '3.8'
torch==2.6.0; python_full_version >= '3.9.0'
torchvision==0.21.0; python_version >= '3.9'
tornado==6.4.2; python_version >= '3.8'
traitlets==5.14.3; python_version >= '3.8'
triton==3.2.0
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
urllib3==2.2.3; python_version >= '3.8'
virtualenv==20.29.1; python_version >= '3.8'
wcwidth==0.2.13
werkzeug==3.1.3; python_version >= '3.9'
wheel==0.45.1; python_version >= '3.8'
wrapt==1.17.0; python_version >= '3.8'
